Rearrange structure
