+++ app/Support/Search/OperatorQuerySearch.php
@@ -39,6 +39,11 @@ use FireflyIII\Repositories\Budget\BudgetRepositoryInterface;
 use FireflyIII\Repositories\Category\CategoryRepositoryInterface;
 use FireflyIII\Repositories\Tag\TagRepositoryInterface;
 use FireflyIII\Repositories\UserGroups\Currency\CurrencyRepositoryInterface;
+use FireflyIII\Support\Search\QueryParser\QueryParserInterface;
+use FireflyIII\Support\Search\QueryParser\Node;
+use FireflyIII\Support\Search\QueryParser\Field;
+use FireflyIII\Support\Search\QueryParser\Word;
+use FireflyIII\Support\Search\QueryParser\Subquery;
 use FireflyIII\Support\ParseDateString;
 use FireflyIII\User;
 use Illuminate\Pagination\LengthAwarePaginator;
