+++ app/Support/Search/QueryParser/GdbotsQueryParser.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace FireflyIII\Support\Search;
+namespace FireflyIII\Support\Search\QueryParser;
 
 use FireflyIII\Exceptions\FireflyException;
 use Gdbots\QueryParser\QueryParser as BaseQueryParser;
